+++ crates/client/src/scan_main.rs
@@ -9,7 +9,7 @@ mod upload;
 use std::path::PathBuf;
 
 use anyhow::{Context, Result};
-use clap::Parser;
+use clap::{CommandFactory, FromArgMatches, Parser};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, Layer};
 
 use find_common::config::{default_config_path, parse_client_config};
@@ -57,7 +57,7 @@ async fn main() -> Result<()> {
         .with(tracing_subscriber::fmt::layer().with_filter(LogIgnoreFilter))
         .init();
 
-    let args = Args::parse();
+    let args = Args::from_arg_matches(&Args::command().version(find_common::tool_version!()).get_matches()).unwrap_or_else(|e| e.exit());
 
     let config_path = args.config.unwrap_or_else(default_config_path);
     let config_str = std::fs::read_to_string(&config_path)