+++ crates/client/src/upload_main.rs
@@ -4,7 +4,7 @@ mod upload;
 use std::path::PathBuf;
 
 use anyhow::{Context, Result};
-use clap::Parser;
+use clap::{CommandFactory, FromArgMatches, Parser};
 
 use find_common::config::{default_config_path, parse_client_config};
 
@@ -40,7 +40,7 @@ async fn main() -> Result<()> {
         )
         .init();
 
-    let args = Args::parse();
+    let args = Args::from_arg_matches(&Args::command().version(find_common::tool_version!()).get_matches()).unwrap_or_else(|e| e.exit());
 
     let config_path = args.config.unwrap_or_else(default_config_path);
     let config_str = std::fs::read_to_string(&config_path)